chore: Update GitService clone_from method to use new repository URL
This commit updates the `clone_from` method in the `GitService` class of the `main.py` file. It modifies the method to use a new repository URL for cloning. The previous URL was replaced with "http://197.255.122.208/djemmanuelk20/model-jaal-net.git". This change ensures that the correct repository is cloned and improves the accuracy of the codebase.