Refactor Dockerfile to use nvidia/cuda base image and install python dependencies
This commit updates the Dockerfile in the deploy directory to use the `nvidia/cuda:12.3.1-base-ubuntu20.04` base image instead of `python:3.11.4-slim-bullseye`. It also installs `python3-pip`, `python3-dev`, and `python-is-python3` packages to ensure the necessary Python dependencies are available. This change improves the compatibility and performance of the Docker image.